Completely unversion libraries in report

+++ packages/protocol/scripts/bash/release-lib.sh
@@ -1,12 +1,12 @@
 #!/usr/bin/env bash
 
 function build_tag() {
-  BRANCH="$1"
-  LOG_FILE="$2"
+  local BRANCH="$1"
+  local LOG_FILE="$2"
 
   echo " - Checkout contracts source code at $BRANCH"
   BUILD_DIR=$(echo build/$(echo $BRANCH | sed -e 's/\//_/g'))
-  rm -r contracts
+  [ -d contracts ] && rm -r contracts
   git checkout $BRANCH -- contracts 2>>$LOG_FILE >> $LOG_FILE
 
   [ -d "build/contracts" ] && mv build/contracts build/contracts_tmp
@@ -20,6 +20,6 @@ function build_tag() {
   fi
   [ -d "build/contracts_tmp" ] && mv build/contracts_tmp build/contracts
 
-  rm -r contracts
+  [ -d contracts ] && rm -r contracts
   git checkout - -- contracts 2>>$LOG_FILE >> $LOG_FILE
 }
